pyproject.toml
[build-system]
# 40.8.0 is the first version compatible with PEP 517
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mag2exp"
version = "0.61.0"
description = "Simulate experimental measurements for magnetic materials"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}

authors = [
    {name = "Samuel Holt"},
    {name = "James Loudon"},
    {name = "Thomas Hicken"},
    {name = "David Cortés Ortuño"},
    {name = "Martin Lang"},
    {name = "Marijan Beg"},
    {name = "Hans Fangohr"}
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Visualization"
]

dependencies = [
    "oommfc~=0.61.0",
    "numpy>=1.19",
    "scipy>=1.6"
]

[project.optional-dependencies]
dev = [
    "build",
    "invoke",
    "nbval",
    "pre-commit",
    "pytest-cov",
    "twine"
]

[project.urls]
homepage = "https://ubermag.github.io"
documentation = "https://ubermag.github.io/documentation/mag2exp"
repository = "https://github.com/ubermag/mag2exp"



# to define additional information we use a custom table
# we need to own the entry for tool.NAME on PyPI
# see: https://www.python.org/dev/peps/pep-0518/#tool-table
[tool.ubermag]
doi = "10.5281/zenodo.3539461"
copyright_holder = "Samuel Holt"
contributors = [
]
about = """
- simulations of experimental techniques perfomed on magnetisation textures,
- the ability to perform "virtual" experiments,
- the functionality to compare of real experimental results agaist simulated experiments.
"""

[tool.coverage.run]
omit = ["mag2exp/tests/*"]

[tool.isort]
profile = "black"
skip_gitignore = true  # ignores files listed in .gitignore

[tool.black]
experimental-string-processing = true
